code/src/gdaip-backend/requirements.txt
fastapi>=0.95.0
uvicorn>=0.21.0
python-multipart>=0.0.6
langchain>=0.0.154
pandas>=1.5.0
scikit-learn>=1.2.0
numpy>=1.23.0
python-dotenv>=1.0.0
pydantic>=1.10.0
requests>=2.28.0
PyPDF2>=3.0.0
pytesseract>=0.3.10
pdf2image>=1.16.3
camelot-py>=0.10.1
opencv-python>=4.6.0  # Required for pdf2image